Update and rename README.md to DataAnalysiswithPython.py
This code loads data from a CSV file, displays the first 5 rows of the data, calculates basic statistics such as the mean, median, and mode, and plots a histogram of the data. It also saves the histogram plot to a file named 'histogram.png'. This is just a simple example of what can be done with data analysis in Python using pandas.